requirements.txt
PyYAML>=6.0.1
confluent-kafka>=2.3.0
numpy>=1.24.0
Pillow>=10.0.0
onnxruntime>=1.17.0
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.31.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.9
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
